feat(feature): update home feature - home feature 의 레이아웃을 업데이트합니다.
